🔨Refactor : price 자료형 변경
Int 타입에서 Long 타입으로 변경

Ref: #8

+++ src/main/kotlin/com/teamsparta/buysell/domain/post/dto/response/PostListResponse.kt
@@ -10,5 +10,5 @@ data class PostListResponse(
     val id: Int,
     val title: String,
     val createdName: String,
-    val price: Int
+    val price: Long
 )
